+++ infrastructures/symfony/Listener/KernelListener.php
@@ -52,7 +52,6 @@ class KernelListener
         private ManagerInterface $manager,
         private ClientWithResponseEventInterface $client,
         private HttpMessageFactoryInterface $factory,
-        private bool $clientInSilentMode = true,
     ) {
     }
 
@@ -82,15 +81,11 @@ class KernelListener
         }
 
         $client = clone $this->client;
-        if (false === $this->clientInSilentMode) {
-            $client->mustSendAResponse();
-        }
-
         $client->setRequestEvent($event);
 
         $psrRequest = $this->getPsrRequest($event->getRequest());
         $this->manager->receiveRequest($client, $psrRequest);
-        $client->sendResponse(null);
+        $client->sendResponse();
 
         return $this;
     }